requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.4=py38hef030d1_0
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38hca72f7f_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38hecd8cb5_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=hb7f2c08_8
brotli-bin=1.0.9=hb7f2c08_8
brotlipy=0.7.0=py38h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.5.7=h8857fd0_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py38h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=py38h98b9b1b_0
cryptography=39.0.1=py38hf6deb26_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=h18a8e69_0
debugpy=1.5.1=py38he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=py_0
entrypoints=0.4=py38hecd8cb5_0
et_xmlfile=1.1.0=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=he9d5cce_0
faiss=1.7.2=py38hcc9e360_2_cpu
ffmpeg=4.3=h0a44026_0
filelock=3.9.0=py38hecd8cb5_0
fonttools=4.39.4=py38hcafd530_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.3.3=py38hef030d1_0
fsspec=2023.5.0=pypi_0
gettext=0.21.0=he85b6c0_1
giflib=5.2.1=h6c40b1e_3
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.31=pyhd8ed1ab_0
glib=2.63.1=hd977a24_0
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py38hd5de756_0
gnutls=3.6.15=hed9c0bf_0
huggingface-hub=0.14.1=pypi_0
icu=58.2=h0a44026_3
idna=3.4=py38hecd8cb5_0
importlib-resources=3.3.1=pyhd8ed1ab_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=hecd8cb5_3538
ipdb=0.13.13=pyhd8ed1ab_0
ipykernel=6.9.1=py38hecd8cb5_0
ipython=8.3.0=py38hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38hecd8cb5_1
jinja2=3.1.2=py38hecd8cb5_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h6c40b1e_1
jsonschema=4.4.0=py38hecd8cb5_0
jupyter=1.0.0=py38_7
jupyter_client=7.2.2=py38hecd8cb5_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py38hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=py38h98b9b1b_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libblas=3.9.0=12_osx64_mkl
libbrotlicommon=1.0.9=hb7f2c08_8
libbrotlidec=1.0.9=hb7f2c08_8
libbrotlienc=1.0.9=hb7f2c08_8
libcblas=3.9.0=12_osx64_mkl
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libfaiss=1.7.2=h88b3cc4_2_cpu
libfaiss-avx2=1.7.2=h1234567_2_cpu
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=11_3_0_h97931a8_31
libgfortran5=12.2.0=he409387_31
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
liblapack=3.9.0=12_osx64_mkl
libpng=1.6.39=h6c40b1e_0
libprotobuf=3.20.3=hfff2838_0
libsodium=1.0.18=h1de35cc_0
libtasn1=4.19.0=h6c40b1e_0
libtiff=4.5.0=hcec6c5f_2
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxml2=2.10.3=h930c0e2_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.1=py38hca72f7f_0
matplotlib=3.7.1=py38h50d1736_0
matplotlib-base=3.7.1=py38hcb346ec_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h1de35cc_1001
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.1=py38h4ab4a9b_0
mkl_random=1.2.2=py38hb2f4e1b_0
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.2.1=py38hecd8cb5_0
multidict=6.0.4=py38hef030d1_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=py38hecd8cb5_0
nbconvert=6.4.4=py38hecd8cb5_0
nbformat=5.3.0=py38hecd8cb5_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.5=py38hecd8cb5_0
nettle=3.7.3=h230ac6f_1
networkx=2.8.4=py38hecd8cb5_1
nltk=3.8.1=pypi_0
notebook=6.4.11=py38hecd8cb5_0
numpy=1.24.3=py38he696674_0
numpy-base=1.24.3=py38h9cd3388_0
openai=0.27.6=pyhd8ed1ab_0
openh264=2.1.1=h8346a28_0
openpyxl=3.1.2=py38hcafd530_0
openssl=1.1.1t=hfd90126_0
packaging=21.3=pyhd3eb1b0_0
pandas=2.0.0=py38hec72209_0
pandas-stubs=2.0.1.230501=pyhd8ed1ab_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=py_1
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py38hcec6c5f_0
pip=23.0.1=py38hecd8cb5_0
platformdirs=3.5.1=pyhd8ed1ab_0
plotly=5.14.1=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.20.3=py38h4cd09af_1
psutil=5.9.5=py38hef030d1_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.0.0=py38hecd8cb5_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h655552a_2
pyrsistent=0.18.0=py38hca72f7f_0
pysocks=1.7.1=py38_1
python=3.8.16=h218abb5_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_0
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py38hef030d1_5
pyzmq=22.3.0=py38he9d5cce_2
qt=5.9.7=h468cd18_1
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.2=hca72f7f_0
regex=2023.5.5=pypi_0
requests=2.29.0=py38hecd8cb5_0
scikit-learn=1.2.2=py38h1081964_1
scipy=1.10.1=py38h1a492ef_2
send2trash=1.8.0=pyhd3eb1b0_1
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.21.1=pyhd8ed1ab_0
setproctitle=1.3.2=py38hef030d1_1
setuptools=66.0.0=py38hecd8cb5_0
sip=4.19.8=py38h0a44026_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py38hecd8cb5_0
tenacity=8.2.2=pyhd8ed1ab_0
terminado=0.13.1=py38hecd8cb5_0
testpath=0.6.0=py38hecd8cb5_0
threadpoolctl=3.1.0=pyh8a188c0_0
tiktoken=0.4.0=pypi_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pyhd8ed1ab_0
torchvision=0.15.2=py38_cpu
tornado=6.1=py38h9ed2024_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.29.2=pypi_0
types-pytz=2023.3.0.0=pyhd8ed1ab_0
typing-extensions=4.5.0=py38hecd8cb5_0
typing_extensions=4.5.0=py38hecd8cb5_0
unicodedata2=15.0.0=py38hef030d1_0
urllib3=1.26.15=py38hecd8cb5_0
wandb=0.15.3=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
wheel=0.38.4=py38hecd8cb5_0
widgetsnbextension=3.5.2=py38hecd8cb5_0
xz=5.4.2=h6c40b1e_0
yaml=0.2.5=h0d85af4_2
yarl=1.9.1=py38hcafd530_0
zeromq=4.3.4=h23ab428_0
zipp=3.8.0=py38hecd8cb5_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
